replace riderstate logic with riders EntityLivingBase#isJumping
The riderstate felt completly overkill or even useless once i realized
that sneaking wont be a thing while riding a chocobo. Also, players
will send their jumping and sneaking state to the server IF they are
riding an entity, which makes this a lot easier. It just needs an
access transformer on the EntityLivingBase#isJumping field since it is
protected by default.