Improved EntityRendererChocobo and Fixed a bug with logging out mounted
Simplified Chocobo Textures to use a Colored base layer with added Layer Renders for Collar, Plumage, and Saddle.

This GREATLY reduces the footprint chococraft makes on a given instances texture map.

Added an event to RidingEventHandler to catch when a player logs out riding a chocobo to forcibly demount them to prevent issues.